+++ docs/.vitepress/config.ts
@@ -12,7 +12,7 @@ import { markdownItImageSize } from 'markdown-it-image-size'
 import { extendConfig } from '@voidzero-dev/vitepress-theme/config'
 import type { FooterLink } from '@voidzero-dev/vitepress-theme'
 import packageJson from '../../packages/vite/package.json' with { type: 'json' }
-import { buildEnd } from './buildEnd.config'
+import { buildEnd } from './buildEnd.config.ts'
 
 const viteVersion = packageJson.version
 const viteMajorVersion = +viteVersion.split('.')[0]