+++ packages/vite/src/node/nativeConfigCompat.ts
@@ -217,7 +217,13 @@ export function createNativeConfigCompatPlugin(
   return {
     name: 'vite:native-config-compat-check',
     transform: {
-      filter: { id: /\.[cm]?[jt]sx?$/ },
+      filter: {
+        id: {
+          include: /\.[cm]?[jt]sx?$/,
+          // exclude rolldown runtime
+          exclude: /^\0/,
+        },
+      },
       async handler(code, id) {
         const isESM =
           typeof process.versions.deno === 'string' || isFilePathESM(id)